backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-socketio>=5.11.0
aiohttp>=3.11.0
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
opinion-clob-sdk>=0.4.3
web3>=6.0.0
